Fix divide

+++ src/divide.c
@@ -4,5 +4,5 @@
 
 double divide(int x, int y)
 {
-    return x / y;
+    return (double)x / y;
 }
